+++ app/Http/Controllers/Admin/JobExperienceController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Job;
 use App\Models\JobExperience;
 use App\Services\Notify;
 use App\Traits\Searchable;
@@ -80,6 +81,13 @@ class JobExperienceController extends Controller
      */
     public function destroy(string $id)
     {
+        // validation
+        $jobExist = Job::where('job_experience_id', $id)->exists();
+
+        if($jobExist) {
+            return response(['message' => 'This item is already been used can\'t delete!'], 500);
+        }
+
         try {
             JobExperience::findOrFail($id)->delete();
             Notify::deletedNotification();
