+++ app/Http/Controllers/Admin/JobTypeController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Job;
 use App\Models\JobType;
 use App\Services\Notify;
 use App\Traits\Searchable;
@@ -81,6 +82,13 @@ class JobTypeController extends Controller
      */
     public function destroy(string $id)
     {
+        // validation
+        $jobExist = Job::where('job_type_id', $id)->exists();
+
+        if($jobExist) {
+            return response(['message' => 'This item is already been used can\'t delete!'], 500);
+        }
+
         try {
             JobType::findOrFail($id)->delete();
             Notify::deletedNotification();
